requirements.txt
python-telegram-bot==12.4.2
google-cloud-firestore==1.6.2
